Several inspirations
I. Use finally to make exception more clean and nice
II. LNK 2001: the static m_Instance in Singleton class should be given a nullptr value at the beginning.
III. LNK 2005: The Vector.h was contained for several times. Thus, it should not contain the definition of an function, just the declaration. Cantious for all head files!
IV. Singleton should be used carefully about the private/public of its construction function and inheritance.
V. The override of operators will be replaced when the class is inherited. It will be covered automatically by the compiler. Therefore, I need to rewrite the override of operators.