Add the bounding box and how to enter bbox
First deal with the deconstruction of singleton. Then add the bounding box as a KD tree with layer divided by x,y,z. Then add the construction to store vertexs and normals. Add the corresponding way to append a rectangle to the world and insert the pointer to the triangle mesh to the bounding box.